Add support for env_dir instead of user-env-compile
user-env-compile has been deprecated and replaced
with a new 3rd env_dir argument. This change updates
the buildpack to extract the GITHUB_AUTH_TOKEN
from the env_dir with a fallback to fetch it from ENV
for apps with user-env-compile still enabled. Once all apps
have been unflagged, the fallback can be removed.

Ref:
https://devcenter.heroku.com/changelog-items/416
https://devcenter.heroku.com/articles/buildpack-api#bin-compile